Moved database related methods to the new Cake\Database\Schema\TableSchemaInterface.
